战斗结算球员状态数据表； 杯数变化；
git-svn-id: svn://ci.ftxgame.local:8199/repos/ArmChairQuatebackDoc/sourcecode/arm_chair_quaterback@646277 11eda05a-b276-4d74-a309-1d70aa31e1c1

+++ lib/common/entities/pk_result_updated_entity.dart
@@ -83,7 +83,7 @@ class PkResultUpdatedAwayTeamResultScoreBoardDetailList {
 	late int ptsPoint = 0;
 	late int reb = 0;
 	late int rebPoint = 0;
-	late int score = 0;
+	late double score = 0;
 	late bool starter = false;
 	late int starterPos = 0;
 	late int stl = 0;
@@ -155,7 +155,7 @@ class PkResultUpdatedAwayTeamResultTeamPlayers {
 	late int playerStatusValue = 0;
 	late int position = 0;
 	late PkResultUpdatedAwayTeamResultTeamPlayersPotential potential;
-	late int power = 0;
+	late double power = 0;
 	late int teamId = 0;
 	late PkResultUpdatedAwayTeamResultTeamPlayersUpStarBase upStarBase;
 	late int updateTime = 0;
@@ -229,6 +229,7 @@ class PkResultUpdatedPlayerResults {
 	late int reb = 0;
 	late int stl = 0;
 	late int teamId = 0;
+	/// 1 mvp 2 reb 3 pts 4 ast
 	late int type = 0;
 
 	PkResultUpdatedPlayerResults();
